refactor: batch 설정 제거 후 quartz 로만 사용
