+++ src/main/java/inno/escape/moviebatch/global/scheduler/BoxOfficeScheduler.java
@@ -1,4 +1,4 @@
-package inno.escape.moviebatch.global.batch.tasklets;
+package inno.escape.moviebatch.global.scheduler;
 
 import inno.escape.moviebatch.domain.boxoffice.dto.daily.DailyBoxOfficeRequestDto;
 import inno.escape.moviebatch.domain.boxoffice.service.BoxOfficeService;
@@ -7,21 +7,19 @@ import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.batch.core.StepContribution;
-import org.springframework.batch.core.scope.context.ChunkContext;
-import org.springframework.batch.core.step.tasklet.Tasklet;
-import org.springframework.batch.repeat.RepeatStatus;
+import org.springframework.scheduling.annotation.Scheduled;
+import org.springframework.stereotype.Service;
 
 @RequiredArgsConstructor
+@Service
 @Slf4j
-public class BoxOfficeTasklet implements Tasklet {
+public class BoxOfficeScheduler implements SchedulerService {
 
   private final BoxOfficeService boxOfficeService;
 
+  @Scheduled(cron = "0 3 0 * * *")
   @Override
-  public RepeatStatus execute(StepContribution contribution, ChunkContext chunkContext) {
-    log.debug("Executed getCompanies Task");
-
+  public void run() {
     String yesterday = DateTimeFormatter.BASIC_ISO_DATE.format(LocalDate.now().minusDays(1));
 
     DailyBoxOfficeRequestDto dto = DailyBoxOfficeRequestDto.builder()
@@ -30,7 +28,13 @@ public class BoxOfficeTasklet implements Tasklet {
         .build();
 
     boxOfficeService.getDailyBoxOffice(dto);
+  }
 
-    return RepeatStatus.FINISHED;
+  @Override
+  public void register() {
+  }
+
+  @Override
+  public void remove() {
   }
 }
